requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asteroid-filterbanks==0.4.0
attrs==25.4.0
av==15.1.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comtypes==1.4.13
contourpy==1.3.3
ctranslate2==4.6.2
cycler==0.12.1
dnspython==2.8.0
docopt==0.6.2
einops==0.8.1
email-validator==2.3.0
fastapi==0.127.1
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
julius==0.2.7
kiwisolver==1.4.9
lazy_loader==0.4
lightning==2.6.0
lightning-utilities==0.15.2
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
mysql-connector-python==8.0.33
narwhals==2.16.0
networkx==3.6.1
nltk==3.9.2
numba==0.64.0
numpy==2.4.2
omegaconf==2.3.0
onnxruntime==1.23.2
optuna==4.6.0
orjson==3.11.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
primePy==1.3
propcache==0.4.1
proto-plus==1.26.1
protobuf==3.20.3
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.11.0
PyMatting==1.1.15
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pypiwin32==223
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
rembg==2.0.72
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
selenium==4.39.0
semver==3.0.4
sentencepiece==0.2.1
sentry-sdk==2.48.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
soundfile==0.13.1
speechbrain==1.0.3
SpeechRecognition==3.14.4
SQLAlchemy==1.4.29
starlette==0.50.0
streamlit==1.54.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2026.3.3
tokenizers==0.22.1
toml==0.10.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchmetrics==1.8.2
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
trio==0.32.0
trio-websocket==0.12.2
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
whisperx==3.7.4
wsproto==1.3.2
yarl==1.22.0
